fix(ralplan): make preflight neutralization atomic
